refactor: rename handler files to match concepts
- workspace-handlers.ts → studio-handlers.ts (git worktree studios)
- workspace-container-handlers.ts → workspace-handlers.ts (org workspaces)
- workspace-container-handlers.test.ts → workspace-handlers.test.ts
- Update all imports in index.ts and test file
