+++ packages/cli/.pcp/runtime/tool-count
@@ -1,0 +1,1 @@
+3
